pyproject.toml
[tool.poetry]
name = "baseball_game_result_notifiy"
version = "0.1.0"
description = ""
authors = ["Haru Nishiura"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
bs4 = "^0.0.2"
ruff = "^0.5.6"
line-bot-sdk = "^3.11.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120